Refactors `registries` into a `WeakMap`.
This is necessary because `CompositeWeakMap.prototype.delete` cannot clean up the `FinalizationRegistry` for the associated composite key (it doesn't know which registry is associated with it and which are not), so the registry leaks. Not only that, but the registry has a strong reference on the composite key it is supposed to clean up, so the composite key leaks as well. By converting this to a `WeakMap`, the registry will automatically clean itself up if the composite key is ever reclaimed, which addresses a potential memory leak.