Refactors `registries` into a `WeakMap`.
This is necessary because `CompositeWeakMap.prototype.delete` cannot clean up registries for the associated composite key, so the registry leaks and the registry has a strong reference on the composite key it is supposed to clean up. By converting this to a `WeakMap`, the registry will automatically clean itself up if the composite key is ever reclaimed, which addresses a potential memory leak.